requirements.txt
affine==2.4.0
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
boto3==1.26.72
botocore==1.29.72
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.13.0
django-extensions==3.2.1
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Fiona==1.9.1
geopandas==0.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
morecantile==3.2.5
munch==2.5.0
numexpr==2.8.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.4.1
pystac==1.6.1
python-dateutil==2.8.2
pytz==2022.7.1
rasterio==1.3.6
requests==2.28.2
rfc3986==1.5.0
rio-cogeo==3.5.0
rio-tiler==4.1.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==2.0.3
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
